Complete LAB03 repository requirements

+++ src/main/java/pl/wsb/fitnesstracker/user/internal/UserServiceImpl.java
@@ -3,7 +3,9 @@ package pl.wsb.fitnesstracker.user.internal;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import pl.wsb.fitnesstracker.user.api.User;
+import pl.wsb.fitnesstracker.user.api.UserNotFoundException;
 import pl.wsb.fitnesstracker.user.api.UserProvider;
 import pl.wsb.fitnesstracker.user.api.UserService;
 
@@ -14,6 +16,7 @@ import java.util.Optional;
 @Service
 @RequiredArgsConstructor
 @Slf4j
+@Transactional
 class UserServiceImpl implements UserService, UserProvider {
 
     private final UserRepository userRepository;
@@ -36,11 +39,7 @@ class UserServiceImpl implements UserService, UserProvider {
     public User updateUser(Long id, User updatedUser) {
 
         User user = userRepository.findById(id)
-                .orElseThrow(() ->
-                        new IllegalArgumentException(
-                                "User with id " + id + " not found"
-                        )
-                );
+                .orElseThrow(() -> new UserNotFoundException(id));
 
         user.update(
                 updatedUser.getFirstName(),
@@ -55,27 +54,42 @@ class UserServiceImpl implements UserService, UserProvider {
     @Override
     public void deleteUser(Long id) {
 
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException(id);
+        }
         userRepository.deleteById(id);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public Optional<User> getUser(final Long userId) {
 
         return userRepository.findById(userId);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public Optional<User> getUserByEmail(final String email) {
 
         return userRepository.findByEmail(email);
     }
 
     @Override
+    @Transactional(readOnly = true)
+    public List<User> findUsersByEmailContaining(String email) {
+
+        return userRepository.findByEmailContainingIgnoreCase(email);
+    }
+
+    @Override
+    @Transactional(readOnly = true)
     public List<User> findAllUsers() {
 
         return userRepository.findAll();
     }
 
+    @Override
+    @Transactional(readOnly = true)
     public List<User> findUsersOlderThan(String date) {
 
         LocalDate localDate = LocalDate.parse(date);
